CarModeReceiver patch
as we say in Italy, "without knowing how to read or write" I can see there are 3 intercepted events, but one of them was not considered in this receiver. IMHO this commit should be cherry picked, bc a static 'isRunning' flag can't work on all setups and a proper stopService() override seems to be proper. I hope I did not misunderstand the author's original intent, IO don't mean to offend anyone :)